minipupper-app/requirements.txt
python-dotenv>=1.0.0
sounddevice>=0.4.5
soundfile>=0.12.1
opencv-contrib-python>=4.8.0
opencv-python-headless>=4.8.0
pyaudio>=0.2.13
numpy>=1.24.0
scipy>=1.11.0

# Cloud Services
google-auth>=2.25.0
google-cloud-texttospeech>=2.14.0
google-cloud-speech>=2.21.0
langchain>=0.1.0
langchain-google-vertexai>=0.1.0
langchain-community>=0.0.10  # For Ollama and other providers

# LLM & Models
ollama>=0.1.0
pydantic>=2.0.0

# Speech Processing
faster-whisper>=0.10.0
pyannote-audio>=3.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
